WIP: Speed optimizations for bearing area calculations; added regression test

+++ cpp/bearing_area.h
@@ -20,10 +20,10 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include "row_major.h"
+#include "eigen_helper.h"
 
 Eigen::ArrayXd nonuniform_bearing_area(Eigen::Ref<Eigen::ArrayXd> topography_x, Eigen::Ref<Eigen::ArrayXd> topography_h,
-                                       Eigen::Ref<Eigen::ArrayXi> topography_s, Eigen::Ref<Eigen::ArrayXd> heights);
+                                       Eigen::Ref<ArrayXl> topography_s, Eigen::Ref<Eigen::ArrayXd> heights);
 Eigen::ArrayXd uniform1d_bearing_area(double dx, Eigen::Ref<Eigen::ArrayXd> topography_h, bool periodic,
                                       Eigen::Ref<Eigen::ArrayXd> heights);
 Eigen::ArrayXd uniform2d_bearing_area(double dx, double dy, Eigen::Ref<RowMajorXXd> topography_h, bool periodic,
